# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Aspectus API
  description: |
    Pandaria 生态统一身份与多租户管理服务。

    ## 认证

    **管理 API** 使用 Service Token：
    ```
    Authorization: Bearer {service_token}
    ```

    **自省端点** 使用 Service Token 认证调用方，subject token 通过 body 传递。

    **Auth 端点** (`/login`、`/register`、`/logout` 等) 无需认证。

    **OAuth2 端点** (`/authorize`, `/oauth/token`) 无需认证。

    ## Token 类型

    | 前缀 | 类型 |
    |------|------|
    | `pk_live_*` | API Key |
    | `eyJ*` | JWT |
    | `ot_*` | Opaque Token |
    | `rt_*` | Refresh Token |

    ## 错误格式

    所有错误使用 [RFC 7807](https://tools.ietf.org/html/rfc7807) Problem Details：
    ```json
    {
      "type": "https://aspectus.dev/errors/unauthorized",
      "title": "Unauthorized",
      "status": 401,
      "detail": "...",
      "instance": "/path"
    }
    ```

    `/introspect` 对无效 token 返回 `{"active": false}` (HTTP 200)，不返回错误。

  version: "0.9.0"
  contact:
    name: Aspectus Team
servers:
  - url: http://localhost:3100
    description: Development server
...
components:
  securitySchemes:
    ServiceToken:
      type: http
      scheme: bearer
      description: Service Token（每个 Project 一个）
  responses:
    NotFound:
      description: 资源不存在
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
...
    ApiKeyListItem:
      type: object
      properties:
        id:
          type: string
        service_account_id:
          type: string
        project:
          type: string
        key_prefix:
          type: string
        scopes:
          type: array
          items:
            type: string
        expires_at:
          type: string
          format: date-time
          nullable: true
        revoked_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
...
paths:
...
  /users/{id}/change-password:
    post:
      tags: [Users]
      summary: 自助修改密码
      description: |
        用户修改自己的登录密码，必须提供当前密码。该端点公开可访问，
        受 IP 级 rate limit 保护。
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
                  minLength: 8
      responses:
        "204":
          description: 修改成功
        "401":
          description: 当前密码错误或用户不存在/被挂起
        "422":
          description: 新密码过短
...
  /api-keys/{id}:
    get:
      tags: [API Keys]
      summary: 查询 API Key
      security:
        - ServiceToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK（不含 key_hash 和原文）
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyListItem"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [API Keys]
      summary: 吊销 API Key
      security:
        - ServiceToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: 吊销成功（幂等）

  # ── Auth (v0.9.0) ──
...